refactor(core): extract session setup into its own module
