+++ src/server/app.js
@@ -28,8 +28,6 @@ import {existsSync, readFileSync} from 'fs';
 
 import BookBrainzData from 'bookbrainz-data';
 import Debug from 'debug';
-import RedisStore from 'connect-redis';
-import {get as _get} from 'lodash';
 import appCleanup from '../common/helpers/appCleanup';
 import compression from 'compression';
 import config from '../common/helpers/config';
@@ -39,10 +37,9 @@ import initInflux from './influx';
 import logNode from 'log-node';
 import logger from 'morgan';
 import path from 'path';
-import redisClient from '../common/helpers/cache';
 import routes from './routes';
 import serveStatic from 'serve-static';
-import session from 'express-session';
+import session from '../common/helpers/session';
 
 
 // Initialize log-to-stdout  writer
@@ -89,28 +86,7 @@ else {
 }
 app.use(express.static(path.join(rootDir, 'static')));
 
-// Set up sessions, using Redis in production and default in-memory for testing environment
-const sessionOptions = {
-	cookie: {
-		maxAge: _get(config, 'session.maxAge', 2592000000),
-		secure: _get(config, 'session.secure', false)
-	},
-	resave: false,
-	saveUninitialized: false,
-	secret: config.session.secret
-};
-if (process.env.NODE_ENV !== 'test') {
-	// eslint-disable-next-line no-console
-	redisClient.connect().catch(redisError => { console.error('Redis error:', redisError); });
-
-	redisClient.on('error', (err) => debug('Redis error:', err));
-
-	const redisStore = new RedisStore({
-		client: redisClient
-	});
-	sessionOptions.store = redisStore;
-}
-app.use(session(sessionOptions));
+app.use(session(process.env.NODE_ENV));
 
 
 if (config.influx) {
